Replace Roboto fonts with Ubuntu fonts (#57)
The Ubuntu fonts generally look better at small point sizes because they
have type hinting for non-antialiased displays. They are the only openly
licensed fonts that I can find with this property.

Not 100% happy with the look of Ubuntu fonts, as they are very wide,
which is not so good for compact displays.

I am keeping the large Roboto fonts in examples, as they look much
better for large point sizes.