Use a common config file for examples. (#80)
This shifts the `init_display` functions in the examples to a
`tempe_config` file that users can replace with something appropriate
for their hardware. It includes example config files for several types
of display.

This also changes the API of the `Display` protocol to require that it
have a `size` attribute than can be used to constrain rendering.

Fixes #68.  Fixes #81.